+++ CHANGELOG.md
@@ -6,9 +6,14 @@ master
 
 Features:
 
+- Data Frame and Expression Filtering #831
 - Allow multiple benchmark paths to be specified from CLI #834
 - Functions which require at least one value return NULL when values are
   empty #835
+
+Improvement:
+
+- Use automatic time unit for expression report #838
 
 1.0.1 (2021-05-11)
 ------------------